requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
PyPDF2==3.0.1
python-docx==1.1.2
tiktoken==0.7.0
hashlib-compat==1.0.0
python-dateutil==2.9.0.post0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
aiosqlite==0.20.0
greenlet==3.1.1
